refactor: move hex as utils::Hex
